some docstrings and refactoring
